Switched RPC based unit syncing to a custom stream to adjust for latency.
Previously, when a unit was ordered by a networked opponent to a location, the networked unit calculated a seperate path in each client, under the assumption this path would be deterministic enough.

That ended up not being the case. I dislike the heavy useage of syncing, but this is currently the best way to fix this issue.

In the future, I'd like to have a dedicated server to be the source of all this truth. For now, its a bizarre mish-mash.